+++ apps/cowswap-frontend/src/modules/tradeFormValidation/pure/TradeFormButtons/tradeButtonsMap.tsx
@@ -1,5 +1,6 @@
 import { ReactElement, ReactNode } from 'react'
 
+import { ACCOUNT_PROXY_LABEL } from '@cowprotocol/common-const'
 import { getIsNativeToken, getWrappedToken } from '@cowprotocol/common-utils'
 import { BridgeProviderQuoteError, BridgeQuoteErrors } from '@cowprotocol/cow-sdk'
 import { CenteredDots, HelpTooltip, TokenSymbol } from '@cowprotocol/ui'
@@ -42,6 +43,10 @@ const quoteErrorTexts: Record<QuoteApiErrorCodes, string> = {
   [QuoteApiErrorCodes.SameBuyAndSellToken]: 'Tokens must be different',
 }
 
+const quoteErrorTextsForBridges: Partial<Record<QuoteApiErrorCodes, string>> = {
+  [QuoteApiErrorCodes.SameBuyAndSellToken]: 'Not yet supported',
+}
+
 const bridgeQuoteErrorTexts: Record<BridgeQuoteErrors, string> = {
   [BridgeQuoteErrors.API_ERROR]: DEFAULT_QUOTE_ERROR,
   [BridgeQuoteErrors.INVALID_BRIDGE]: DEFAULT_QUOTE_ERROR,
@@ -51,6 +56,11 @@ const bridgeQuoteErrorTexts: Record<BridgeQuoteErrors, string> = {
   [BridgeQuoteErrors.NO_INTERMEDIATE_TOKENS]: 'No routes found',
   [BridgeQuoteErrors.NO_ROUTES]: 'No routes found',
   [BridgeQuoteErrors.ONLY_SELL_ORDER_SUPPORTED]: 'Only "sell" orders are supported',
+}
+
+const errorTooltipContentForBridges: Partial<Record<QuoteApiErrorCodes, string>> = {
+  [QuoteApiErrorCodes.SameBuyAndSellToken]:
+    'Bridging without swapping is not yet supported. Let us know if you want this feature!',
 }
 
 // TODO: Add proper return type annotation
@@ -105,15 +115,24 @@ export const tradeButtonsMap: Record<TradeFormValidation, ButtonErrorConfig | Bu
     const { quote } = context
 
     if (quote.error instanceof QuoteApiError) {
-      if (quote.error.type === QuoteApiErrorCodes.UnsupportedToken) {
+      const errorType = quote.error.type
+
+      if (errorType === QuoteApiErrorCodes.UnsupportedToken) {
         return unsupportedTokenButton(context)
       }
 
-      const errorText = quoteErrorTexts[quote.error.type] || DEFAULT_QUOTE_ERROR
+      const isBridge = quote.isBridgeQuote
+      const errorText =
+        (isBridge && quoteErrorTextsForBridges[errorType]) || quoteErrorTexts[errorType] || DEFAULT_QUOTE_ERROR
+
+      const errorTooltipText = isBridge && errorTooltipContentForBridges[errorType]
 
       return (
         <TradeFormBlankButton disabled={true}>
-          <Trans>{errorText}</Trans>
+          <>
+            <Trans>{errorText}</Trans>
+            {errorTooltipText && <HelpTooltip text={errorTooltipText} />}
+          </>
         </TradeFormBlankButton>
       )
     }
@@ -228,12 +247,12 @@ export const tradeButtonsMap: Record<TradeFormValidation, ButtonErrorConfig | Bu
   [TradeFormValidation.ProxyAccountLoading]: {
     text: (
       <>
-        <span>Loading account proxy</span>
+        <span>Loading {ACCOUNT_PROXY_LABEL}</span>
         <CenteredDots smaller />
       </>
     ),
   },
   [TradeFormValidation.ProxyAccountUnknown]: {
-    text: "Couldn't verify account proxy, please try later",
+    text: `Couldn't verify ${ACCOUNT_PROXY_LABEL}, please try later`,
   },
 }