+++ apps/explorer/src/components/orders/BridgeDetailsTable/BridgeDetailsContent/index.tsx
@@ -1,5 +1,6 @@
 import React, { ReactNode } from 'react'
 
+import { RECEIVED_LABEL } from '@cowprotocol/common-const'
 import { BridgeStatus, CrossChainOrder } from '@cowprotocol/cow-sdk'
 
 import { AddressLink } from 'components/common/AddressLink'
@@ -66,7 +67,7 @@ export function BridgeDetailsContent({ crossChainOrder }: BridgeDetailsContentPr
         <AmountSectionWrapper>
           <BridgeAmountDisplay labelPrefix="From:" bridgeToken={sourceToken} amount={inputAmount.toString()} />
           <BridgeAmountDisplay
-            labelPrefix="To at least:"
+            labelPrefix="To:"
             bridgeToken={destinationToken}
             amount={outputAmount?.toString() || '0'}
             bridgeProvider={bridgeProvider}
@@ -74,7 +75,7 @@ export function BridgeDetailsContent({ crossChainOrder }: BridgeDetailsContentPr
         </AmountSectionWrapper>
       </DetailRow>
 
-      <DetailRow label="You received" tooltipText={BridgeDetailsTooltips.youReceived}>
+      <DetailRow label={RECEIVED_LABEL} tooltipText={BridgeDetailsTooltips.youReceived}>
         {outputAmount && destinationToken && bridgeStatus === BridgeStatus.EXECUTED && (
           <BridgeReceiveAmount
             amount={outputAmount}