+++ apps/explorer/src/components/orders/BridgeDetailsTable/bridgeDetailsTooltips.tsx
@@ -1,3 +1,5 @@
+import { ACCOUNT_PROXY_LABEL } from '@cowprotocol/common-const'
+
 import { AccountProxyLink } from '../shared/AccountProxyLink'
 
 export const BridgeDetailsTooltips = {
@@ -11,8 +13,8 @@ export const BridgeDetailsTooltips = {
   provider: 'The bridging solution provider.',
   accountFromProxy: (
     <span>
-      The <AccountProxyLink>Account Proxy</AccountProxyLink> address which will/did receive bought amount of the Swap
-      before the bridge is initiated.
+      The <AccountProxyLink>{ACCOUNT_PROXY_LABEL}</AccountProxyLink> address which will/did receive bought amount of the
+      Swap before the bridge is initiated.
     </span>
   ),
   receiverAddress: 'The account address to which the tokens are bridged on the destination chain.',